MEDIUM: add general cluster controller information to dataplane
When interacting with a dataplane to configure connecting to a controller,
displaying general controller information like name and address is
useful for identifying to which controller this particular datdataplane
is connected to.

Allowing write to these settings can allow to reconfigure connecting to
a different controller, but also to a controller to send updates about
its own information (name, description).